Handle event propagation stopping
