+++ luda-editor/client/src/pages/sequence_edit_page/components/auto_complete_text_input/mod.rs
@@ -23,10 +23,10 @@ pub enum Request {
     Blur,
 }
 
-pub enum Event {
+pub enum Event<'a> {
     TextChange { text: String },
     EditDone,
-    KeyDown { event: KeyboardEvent },
+    KeyDown { event: KeyboardEvent<'a> },
     ReqQueuePopFront,
 }
