+++ namui/src/namui/hooks/ctx/render_ctx/method_impl/effect.rs
