+++ namui/src/namui/hooks/macros.rs
@@ -1,0 +1,12 @@
+/// callback!('a, A)
+#[macro_export]
+macro_rules! callback {
+    ($lifetime: lifetime, $param: ty) => {
+        Box<dyn $lifetime + FnOnce($param)>
+    };
+    ($lifetime: lifetime) => {
+        Box<dyn $lifetime + FnOnce()>
+    };
+}
+
+pub use callback;
